fix: async import init func
