+++ designer-demo/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@opentiny/tiny-engine": "workspace:^",
+    "@opentiny/tiny-engine-meta-register": "^2.3.0",
     "@opentiny/tiny-engine-utils": "workspace:*",
     "@opentiny/vue": "~3.20.0",
     "@opentiny/vue-design-smb": "~3.20.0",
