+++ designer-demo/registry.js
@@ -17,5 +17,19 @@ export default {
   [META_SERVICE.Http]: HttpService,
   'engine.config': {
     ...engineConfig,
+  },
+  'engine.plugins.i18n': {
+    overwrite: {
+      lifeCycles: {
+        '': {
+          onMounted: [
+            (ctx) => () => {
+              console.log('overWrite i18n onMounted', ctx.i18nSearchTypes, ctx.currentSearchType.value)
+              ctx.currentSearchType.value = ctx.i18nSearchTypes[0].value
+            }
+          ]
+        }
+      }
+    }
   }
 }
