+++ designer-demo/src/main.js
@@ -9,15 +9,21 @@
  * A PARTICULAR PURPOSE. SEE THE APPLICABLE LICENSES FOR MORE DETAILS.
  *
  */
-
-// 导入@opentiny/tiny-engine时，内部的依赖包也会逐个导入，可能会执行useComplie，此时需要templateHashMap。所以需要先执行一次defineEntry
-import { init } from '@opentiny/tiny-engine'
+import { defineEntry } from '@opentiny/tiny-engine-meta-register'
 import { configurators } from './configurators/'
 import registry  from '../registry'
 import 'virtual:svg-icons-register'
 
-init({
-  registry,
-  configurators,
-  createAppSignal: ['global_service_init_finish']
-})
+async function startApp() {
+  const { init } = await import('@opentiny/tiny-engine')
+  // 导入@opentiny/tiny-engine时，内部的依赖包也会逐个导入，可能会执行useComplie，此时需要templateHashMap。所以需要先执行一次defineEntry
+  defineEntry(registry)
+
+  init({
+    registry,
+    configurators,
+    createAppSignal: ['global_service_init_finish']
+  })
+}
+
+startApp()
