+++ packages/plugins/materials/src/meta/block/meta.js
@@ -1,4 +1,3 @@
 export default {
-  id: 'engine.plugins.materials.block',
-  type: 'plugins'
+  id: 'engine.plugins.materials.block'
 }
