+++ packages/plugins/materials/src/meta/component/meta.js
@@ -1,4 +1,3 @@
 export default {
-  id: 'engine.plugins.materials.component',
-  type: 'plugins'
+  id: 'engine.plugins.materials.component'
 }
