+++ packages/plugins/materials/src/meta/layout/meta.js
@@ -1,4 +1,3 @@
 export default {
-  id: 'engine.plugins.materials.layout',
-  type: 'plugins'
+  id: 'engine.plugins.materials.layout'
 }
