+++ packages/register/src/layoutHash.ts
@@ -10,9 +10,9 @@
  *
  */
 
-import { layoutHashMap } from './common'
+// import { layoutHashMap } from './common'
 
-export const getLayoutComponent = (metaData: Record<string, any> & { id: string }) => {
-  const customLayout = layoutHashMap[metaData.id]
-  return customLayout
-}
+// export const getLayoutComponent = (metaData: Record<string, any> & { id: string }) => {
+//   const customLayout = layoutHashMap[metaData.id]
+//   return customLayout
+// }
